backend/requirements.txt
blinker==1.6
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
crypto==1.4.1
Flask==2.2.5
Flask-Cors==4.0.0
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Naked==0.1.32
packaging==23.2
Pillow==9.5.0
pycryptodome==3.19.0
PyYAML==6.0.1
requests==2.31.0
shellescape==3.8.1
urllib3==2.0.7
Werkzeug==2.2.3